Add a new story for ProductCard
